backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
tensorflow
firebase-admin
